Made contradicting cla's mutually exclusive.
For the command-line arguments of clang-build one can specify to build "ALL" targets, meaning those targets not needed by the top-level project are still selected to be build. Or one can select those targets one wishes to build (and all their dependencies). It doesn't make sense to provide both. With this patch, the user will be informed that they provided contradicting arguments.