Initial compile-commands implementation
`Toolchain` now returns also the depfile and compilation commands, which
are used by `SingleSource`. It is checked whether an entry in the
database, if it exists, already contains the same file and output.

The current implementation is lacking: the database is read in twice for
each source file, code is partially duplicated and it might be preferable
to update the compilation database _before_ actually executing each
command.