Emscripten toolchain test
Update required Python version to 3.9 to enable asyncio usage.

Fixed the CI if-statements and added emscripten-installation.

Fixed the `--tolchain` CLI argument passing in the emscripten toolchain
test and added checks whether the expected output was produced.

Also fixed the library output directory of the emscripten toolchain on
Windows, as it doesn't make sense to put wasm files in the bin folder.