Minor fix to depfile parsing
On Windows, it could happen that the first line started with a full path
string, i.e. the first colon would be part of the name of the dependent
file, starting with e.g. "C:\".

The correct way to jumpt to behind the dependent file name is to look
for the first colon with a space behind it.